Bug fixes -> - Compare section graphs missing model names - Alternate method graphs not loading , Feature : >- Add GUI interface for input
