+++ toolkit/analysis_and_visualization/classification/.gitignore
@@ -7,3 +7,4 @@ node_modules
 .stylelintrc.json
 package.json
 package-lock.json
+*.pyc
