+++ toolkit/analysis_and_visualization/classification/assets/templates/index.html
@@ -627,7 +627,7 @@
         </div>
     </div>
 
-
+    <br>
     <div class="row" id="compare-result-graphs">
 
     </div>
